Move index tests to separate files
