+++ ydb/core/kqp/ut/indexes/ya.make
@@ -13,6 +13,8 @@ ENDIF()
 SRCS(
     kqp_indexes_ut.cpp
     kqp_indexes_multishard_ut.cpp
+    kqp_indexes_prefixed_vector_ut.cpp
+    kqp_indexes_vector_ut.cpp
 )
 
 PEERDIR(
